requirements.txt
aiomysql==0.1.1
aioredis==2.0.1
anyio==3.6.2
async-timeout==4.0.2
greenlet==2.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
priority==2.0.0
pydantic==1.10.2
PyJWT==2.6.0
PyMySQL==1.0.2
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.21.0
starlette-context==0.3.4
toml==0.10.2
typing_extensions==4.4.0
uvloop==0.17.0
wsproto==1.2.0
